fix c89 compilation error in r4748
additional in libcob:
* intrinsic.c (cob_intr_bit_of, cob_intr_bit_to_char,cob_intr_hex_of cob_intr_hex_to_char): fixed invalid attribute of return field

+++ libcob/strings.c
@@ -878,7 +878,7 @@ cob_unstring_into (cob_field *dst, cob_field *dlm, cob_field *cnt)
 		return;
 	}
 
-	if (unstring_offset >= unstring_src->size) {
+	if (unstring_offset >= (int)unstring_src->size) {
 		/* overflow from the last iteration (multiple INTO targets) */
 		return;
 	}
